separate asset and library headers
