+++ headers/enduro2d/high/_all.hpp
@@ -30,6 +30,8 @@
 #include "systems/render_system.hpp"
 
 #include "address.hpp"
+#include "asset.hpp"
+#include "asset.inl"
 #include "atlas.hpp"
 #include "library.hpp"
 #include "library.inl"
